M2_detect/requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.1.1
cycler==0.12.1
filelock==3.16.1
fonttools==4.57.0
fsspec==2025.3.0
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
opencv-python==4.12.0.88
packaging==25.0
pandas==2.0.3
pillow==10.4.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
spiceypy==6.0.0
sympy==1.13.3
threadpoolctl==3.5.0
torch==2.4.1
torchvision==0.19.1
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.186
ultralytics-thop==2.0.16
urllib3==2.2.3
zipp==3.20.2
